authorization not working but impressive(will be working in next commit i think)

+++ src/main/java/pl/mazy/todoapp/requests/NoteReq.java
@@ -21,7 +21,7 @@ public class NoteReq {
     }
 
     record newNoteReq(
-            Integer owner_id,
+            Long owner_id,
             String name,
             String description
     ){ }
